requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
cloudinary==1.38.0
cryptography==42.0.1
dj-database-url==2.1.0
Django==5.0.1
django-daraja==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
idna==3.6
numpy==1.26.3
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
pypng==0.20220715.0
python-decouple==3.8
pytz==2023.4
pyzbar==0.1.9
qrcode==7.4.2
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.1.0
